Move handlers and httpd from cmd/services/m3coordinator to src/coordinator
