+++ src/coordinator/handler/namespace/add_test.go
