+++ src/coordinator/handler/namespace/common_test.go
