+++ src/coordinator/handler/namespace/delete.go
@@ -28,7 +28,7 @@ import (
 	"strings"
 
 	clusterclient "github.com/m3db/m3cluster/client"
-	"github.com/m3db/m3db/src/cmd/services/m3coordinator/handler"
+	"github.com/m3db/m3db/src/coordinator/handler"
 	"github.com/m3db/m3db/src/coordinator/util/logging"
 
 	"github.com/m3db/m3db/src/dbnode/storage/namespace"
