+++ src/coordinator/handler/placement/add_test.go
