+++ src/coordinator/handler/placement/common_test.go
