+++ src/coordinator/handler/placement/delete_all_test.go
