+++ src/coordinator/handler/placement/delete_test.go
