+++ src/coordinator/handler/prometheus/remote/read.go
@@ -25,8 +25,8 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/m3db/m3db/src/cmd/services/m3coordinator/handler"
-	"github.com/m3db/m3db/src/cmd/services/m3coordinator/handler/prometheus"
+	"github.com/m3db/m3db/src/coordinator/handler"
+	"github.com/m3db/m3db/src/coordinator/handler/prometheus"
 	"github.com/m3db/m3db/src/coordinator/executor"
 	"github.com/m3db/m3db/src/coordinator/generated/proto/prompb"
 	"github.com/m3db/m3db/src/coordinator/storage"
