+++ src/coordinator/handler/prometheus/remote/write.go
@@ -24,8 +24,8 @@ import (
 	"context"
 	"net/http"
 
-	"github.com/m3db/m3db/src/cmd/services/m3coordinator/handler"
-	"github.com/m3db/m3db/src/cmd/services/m3coordinator/handler/prometheus"
+	"github.com/m3db/m3db/src/coordinator/handler"
+	"github.com/m3db/m3db/src/coordinator/handler/prometheus"
 	"github.com/m3db/m3db/src/coordinator/generated/proto/prompb"
 	"github.com/m3db/m3db/src/coordinator/storage"
 	"github.com/m3db/m3db/src/coordinator/util/execution"
