+++ src/coordinator/httpd/handler.go
@@ -26,12 +26,12 @@ import (
 	"os"
 
 	"github.com/m3db/m3db/src/cmd/services/m3coordinator/config"
-	"github.com/m3db/m3db/src/cmd/services/m3coordinator/handler"
-	"github.com/m3db/m3db/src/cmd/services/m3coordinator/handler/namespace"
-	"github.com/m3db/m3db/src/cmd/services/m3coordinator/handler/openapi"
-	"github.com/m3db/m3db/src/cmd/services/m3coordinator/handler/placement"
-	"github.com/m3db/m3db/src/cmd/services/m3coordinator/handler/prometheus/native"
-	"github.com/m3db/m3db/src/cmd/services/m3coordinator/handler/prometheus/remote"
+	"github.com/m3db/m3db/src/coordinator/handler"
+	"github.com/m3db/m3db/src/coordinator/handler/namespace"
+	"github.com/m3db/m3db/src/coordinator/handler/openapi"
+	"github.com/m3db/m3db/src/coordinator/handler/placement"
+	"github.com/m3db/m3db/src/coordinator/handler/prometheus/native"
+	"github.com/m3db/m3db/src/coordinator/handler/prometheus/remote"
 	"github.com/m3db/m3db/src/coordinator/executor"
 	"github.com/m3db/m3db/src/coordinator/storage"
 	"github.com/m3db/m3db/src/coordinator/util/logging"
